+++ packages/react-cookie/src/types.ts
@@ -1,6 +1,8 @@
-import Cookies from 'universal-cookie';
+import Cookies, { Cookie } from 'universal-cookie';
 
 export type ReactCookieProps = {
   cookies?: Cookies;
+  allCookies?: { [name: string]: Cookie };
   children?: any;
+  ref?: React.RefObject<{}>;
 };